:recycle: [refactory]: Add Component of UserItem
- Add UserItem.tsx
- Add Total count in UserList.tsx
