+++ src/pages/users/page/UserBody/UserList/index.ts
@@ -1,0 +1,1 @@
+export { default as UserList } from './UserList';
